feat: drop 1.21.11 & bump sbapi
