+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -26,7 +26,7 @@ body:
         attributes:
             label: SkyBlockPv Version
             description: The version of SkyBlockPv you are using.
-            placeholder: e.g., "1.8.2-1.21.11"
+            placeholder: e.g., "1.8.4-26.1"
         validations:
             required: true
     -   type: input
