+++ .github/workflows/gradle.yml
@@ -26,8 +26,4 @@ jobs:
             -   uses: actions/upload-artifact@v7.0.0
                 with:
                     archive: false
-                    path: build/libs/*-1.21.11.jar
-            -   uses: actions/upload-artifact@v7.0.0
-                with:
-                    archive: false
                     path: build/libs/*-26.1.jar
