--- build.obf.gradle.kts
@@ -1,15 +1,0 @@
-@file:Suppress("UnstableApiUsage")
-
-plugins {
-    id("net.fabricmc.fabric-loom-remap")
-    `pv-setup`
-}
-
-dependencies {
-    mappings(loom.layered {
-        officialMojangMappings()
-        parchment(variantOf(versionedCatalog["parchment"]) {
-            artifactType("zip")
-        })
-    })
-}
