--- gradle/1_21_11.versions.toml
@@ -1,23 +1,0 @@
-[versions]
-minecraft-start = "1.21.11"
-minecraft-end = "1.21.11"
-minecraft = "1.21.11"
-parchment = "2025.10.05"
-
-fabric-api = "0.140.0+1.21.11"
-
-resourceful-lib = "3.11.0"
-resourceful-config = "3.11.0"
-olympus = "1.7.0"
-placeholders = "2.8.1+1.21.10"
-
-[libraries]
-minecraft = { module = "com.mojang:minecraft", version.ref = "minecraft" }
-
-parchment = { module = "org.parchmentmc.data:parchment-1.21.9", version.ref = "parchment" }
-fabric-api = { module = "net.fabricmc.fabric-api:fabric-api", version.ref = "fabric-api" }
-
-resourceful-lib = { module = "com.teamresourceful.resourcefullib:resourcefullib-fabric-1.21.11", version.ref = "resourceful-lib" }
-resourceful-config = { module = "com.teamresourceful.resourcefulconfig:resourcefulconfig-fabric-1.21.11", version.ref = "resourceful-config" }
-olympus = { module = "earth.terrarium.olympus:olympus-fabric-1.21.11", version.ref = "olympus" }
-placeholders = { module = "eu.pb4:placeholder-api", version.ref = "placeholders" }
