+++ gradle/libs.versions.toml
@@ -10,7 +10,7 @@ kotlin-gradle-plugin = "2.4.0"
 kotlin-symbol-processor = "2.3.5"
 
 # Dependencies
-skyblockapi = "4.1.11"
+skyblockapi = "4.2.2"
 repolib = "2.0.2"
 
 resourceful-configkt = "3.5.12"
@@ -18,7 +18,7 @@ resourceful-configkt = "3.5.12"
 hypixelapi = "1.0.1+build.1+mc1.21"
 mixinconstraints = "1.0.7"
 
-meowdding-lib = "4.1.7"
+meowdding-lib = "4.2.0-beta.1"
 meowdding-ktmodules = "1.0.5"
 meowdding-ktcodecs = "1.0.28"
 meowdding-resources = "1.0.17"
