--- versions/1.21.11/src/main/java/me/owdding/skyblockpv/mixin/AvatarRenderStateMixin.java
@@ -1,60 +1,0 @@
-package me.owdding.skyblockpv.mixin;
-
-import me.owdding.lib.rendering.text.TextShader;
-import me.owdding.skyblockpv.utils.PlayerRenderStateAccessor;
-import net.minecraft.client.renderer.entity.state.AvatarRenderState;
-import org.jetbrains.annotations.Nullable;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
-
-import java.net.URI;
-
-@Mixin(AvatarRenderState.class)
-public class AvatarRenderStateMixin implements PlayerRenderStateAccessor {
-
-    @Unique
-    private URI skyblockpv$catOnShould = null;
-    @Unique
-    private TextShader skyblockpv$nameShader = null;
-    @Unique
-    private TextShader skyblockpv$scoreShader = null;
-
-    @Override
-    public @Nullable URI getSkyblockpv$catOnShoulder() {
-        return skyblockpv$catOnShould;
-    }
-
-    @Override
-    public void setSkyblockpv$catOnShoulder(@Nullable URI catOnShoulder) {
-        skyblockpv$catOnShould = catOnShoulder;
-    }
-
-    @Override
-    public @Nullable TextShader getSkyblockpv$nameShader() {
-        return this.skyblockpv$nameShader;
-    }
-
-    @Override
-    public void setSkyblockpv$nameShader(@Nullable TextShader textShader) {
-        this.skyblockpv$nameShader = textShader;
-    }
-
-    @Override
-    public @Nullable TextShader getSkyblockpv$scoreShader() {
-        return this.skyblockpv$scoreShader;
-    }
-
-    @Override
-    public void setSkyblockpv$scoreShader(@Nullable TextShader textShader) {
-        this.skyblockpv$scoreShader = textShader;
-    }
-
-    @Override
-    public @Nullable Boolean getSkyblockpv$isCatBaby() {
-        return null;
-    }
-
-    @Override
-    public void setSkyblockpv$isCatBaby(@Nullable Boolean ignored) {
-    }
-}
