--- versions/1.21.11/src/main/resources/assets/skyblock-pv/shaders/core/inventory.vsh
@@ -1,18 +1,0 @@
-#version 150
-
-//!moj_import <minecraft:dynamictransforms.glsl>
-//!moj_import <minecraft:projection.glsl>
-
-out vec2 texCoord0;
-out vec4 vertexColor;
-
-in vec3 Position;
-in vec2 UV0;
-in vec4 Color;
-
-void main() {
-    gl_Position = ProjMat * ModelViewMat * vec4(Position, 1.0);
-
-    texCoord0 = UV0;
-    vertexColor = Color;
-}
