--- versions/1.21.11/src/main/resources/assets/skyblock-pv/shaders/text/gradient.vsh
@@ -1,27 +1,0 @@
-#version 150
-
-//!moj_import <minecraft:dynamictransforms.glsl>
-//!moj_import <minecraft:projection.glsl>
-//!moj_import <minecraft:fog.glsl>
-
-in vec3 Position;
-in vec4 Color;
-in vec2 UV0;
-in ivec2 UV2;
-
-uniform sampler2D Sampler0;
-
-out float sphericalVertexDistance;
-out float cylindricalVertexDistance;
-out vec4 vertexColor;
-out vec2 texCoord0;
-
-void main() {
-    gl_Position = ProjMat * ModelViewMat * vec4(Position, 1.0);
-
-
-    sphericalVertexDistance = fog_spherical_distance(Position);
-    cylindricalVertexDistance = fog_cylindrical_distance(Position);
-    vertexColor = Color * texture(Sampler0, UV2);
-    texCoord0 = UV0;
-}
